検索用のModelを作成
