+++ backend/app/repositories/models/conversation.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import json
+import logging
 import re
 from pathlib import Path
 from typing import TYPE_CHECKING, Annotated, Any, Literal, Self, TypeGuard
@@ -41,6 +42,9 @@ from pydantic import BaseModel, Discriminator, Field, JsonValue, field_validator
 
 if TYPE_CHECKING:
     from app.agents.tools.agent_tool import ToolRunResult
+
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
 
 
 class TextContentModel(BaseModel):
@@ -700,13 +704,6 @@ class ConversationModel(BaseModel):
     should_continue: bool
 
 
-class SearchHighlight(BaseModel):
-    """Model representing highlight information for search results"""
-
-    field_name: str  # "Title" or "MessageMap"
-    fragments: list[str]  # Text fragments containing the search term
-
-
 class ConversationMeta(BaseModel):
     id: str
     title: str
@@ -714,94 +711,6 @@ class ConversationMeta(BaseModel):
     last_updated_time: float = 0.0
     model: str
     bot_id: str | None
-    highlights: list[SearchHighlight] | None = None
-    highlight_texts: list[str] | None = None
-
-    @classmethod
-    def from_opensearch_response(cls, hit: dict) -> Self:
-        """Create a ConversationMeta instance from OpenSearch response"""
-        DEFAULT_MODEL = "claude-v3.5-haiku"
-        source = hit["_source"]
-
-        # Extract conversation ID from SK (e.g. "{user_id}#CONV#{conversation_id}" -> "{conversation_id}")
-        sk = source.get("SK", "")
-        conversation_id = decompose_conv_id(sk)
-
-        # Get model directly from extractedModel field
-        model = source.get("extractedModel", DEFAULT_MODEL)
-
-        # Get create time from source
-        create_time = 0.0
-        if source.get("CreateTime"):
-            create_time = float(source.get("CreateTime"))
-        elif source.get("create_time"):
-            create_time = float(source.get("create_time"))
-
-        # Get last updated time from latest message's create_time
-        last_updated_time = 0.0
-
-        # 1. Try to get from the latest message's create_time
-        if source.get("messages") and isinstance(source.get("messages"), list):
-            messages = source.get("messages", [])
-            message_times = [
-                msg.get("value", {}).get("create_time", 0)
-                for msg in messages
-                if isinstance(msg.get("value"), dict)
-            ]
-            if message_times:
-                last_updated_time = float(max(message_times))
-
-        # 2. Try to get from sort value (timestamp from messages)
-        elif (
-            "sort" in hit
-            and len(hit["sort"]) > 1
-            and isinstance(hit["sort"][1], (int, float))
-        ):
-            last_updated_time = float(hit["sort"][1])
-
-        # If create_time is still 0, use last_updated_time as create_time
-        if create_time == 0.0 and last_updated_time > 0.0:
-            create_time = last_updated_time
-
-        # Create conversation meta instance
-        conversation = cls(
-            id=conversation_id,
-            title=source.get("Title", "Untitled conversation"),
-            create_time=create_time,
-            last_updated_time=last_updated_time,
-            model=model,
-            bot_id=source.get("BotId"),
-        )
-
-        # Add highlight information if available
-        if "highlight" in hit:
-            highlights = []
-            highlight_texts = []
-
-            for field, fragments in hit["highlight"].items():
-                if field == "extractedContent":
-                    highlights.append(
-                        SearchHighlight(field_name="MessageBody", fragments=fragments)
-                    )
-                    highlight_texts.extend(fragments)
-                elif field == "Title":
-                    highlights.append(
-                        SearchHighlight(field_name=field, fragments=fragments)
-                    )
-                    highlight_texts.extend(fragments)
-                elif field == "messages.value.content.body":
-                    highlights.append(
-                        SearchHighlight(field_name="MessageBody", fragments=fragments)
-                    )
-                    highlight_texts.extend(fragments)
-
-            if highlights:
-                conversation.highlights = highlights
-
-            if highlight_texts:
-                conversation.highlight_texts = highlight_texts
-
-        return conversation
 
 
 class RelatedDocumentModel(BaseModel):
