+++ backend/app/usecases/chat.py
@@ -504,7 +504,6 @@ def chat_output_from_message(
     return ChatOutput(
         conversation_id=conversation.id,
         create_time=conversation.create_time,
-        last_updated_time=conversation.last_updated_time,
         message=MessageOutput(
             role=message.role,
             content=[c.to_content() for c in message.content],
@@ -647,7 +646,6 @@ def fetch_conversation(user_id: str, conversation_id: str) -> Conversation:
         id=conversation_id,
         title=conversation.title,
         create_time=conversation.create_time,
-        last_updated_time=conversation.last_updated_time,
         last_message_id=conversation.last_message_id,
         message_map=message_map,
         bot_id=conversation.bot_id,
